Change briefs for access, excess, and variations
"acc-" briefs are generally stroked "A/K…" or sometimes "SE…" while
"exce-" briefs are generally stroked "KPE…", so "KPES" should be
"excess" rather than "access". Similarly, words with a double 's' are
often stroked with "SZ" so we'll create a new brief for "access" using
"SESZ". We'll also include briefs that use suffix strokes like
"SESZ/-BL" for "accessible". The main briefs are now:

- "SESZ": "access"
- "SEBL": "accessible",
- "SEBLT": "accessibility"
- "KPES": "excess"

The "SES/PAOL" brief for "cesspool" should be used instead of
"SESZ/PAOL", so we'll remove the latter to avoid word boundaries errors
when trying to write "access pool" e.g. "To access pool facilities…",
which would otherwise produce "To cesspool facilities…".