Reorganise outlines for expose/suppose and variations
Phonetically, "SPOES" and "SPOEF" could both be "suppose". One of these
could be a brief for "expose", saving us from using "KPOEF", which is
already occupied by "compose", and avoiding asterisks. As "exposedly" is
not a word and "SPOELDS" can't be stroked, we'll make "SPOEF" the
outline for "suppose" and "SPOES" the outline for "expose".

We'll also remove a bunch of redundant briefs.