Update strokes for Israel*
Given the suffix entry: "EUS": "{^is}",

… removing entries starting with "EUS/" for the start of a word will
avoid word boundary errors. For example, if you wanted to write, "… axe
Israel…" using "ABGS EUS/RAEL", this would be translated as "axis real"
unless "EUS/RAEL" is replaced by "EUZ/RAEL", which is already present.

Add condensed strokes for Israelites, Israeli, and Israelis using
suffixes strokes.

It looks like "Israelies" is a misspelling and doesn't need to be
replaced.

It looks like "israelii" is a medical term that we'll keep.