Move condensed stroke -GT for thing to di-briefs
While this is a condensed stroke that already works with existing Plover
dictionary, it doesn't appear to follow any existing convention for
phrasing briefs ending in GT and " thing". As it is shorter than the
current preferred phonetic entry in the top X words dictionaries,
Typey Type would prioritise the shorter entry.