Re-generate emoji.json with new dicts + misstrokes
This is the result of re-generating emoji.json using an updated
stemoji.js script from Typey Type CLI using the following dictionaries
plus the `misstrokes.json` dictionary:

- "abbreviations.json",
- "briefs.json",
- "condensed-strokes-fingerspelled.json",
- "condensed-strokes.json",
- "currency.json",
- "dict.json",
- "dict-en-AU-phonetic.json",
- "dict-en-AU-vocab.json",
- "dict-en-AU-with-extra-stroke.json",
- "nouns.json",
- "numbers.json",
- "numbers-powerups.json",
- "proper-nouns.json",
- "punctuation-powerups.json",
- "punctuation-unspaced.json",
- "symbols.json",
- "symbols-briefs.json",
- "symbols-currency.json",
- "top-10000-project-gutenberg-words.json",
- "fingerspelling.json",
- "top-level-domains.json"

One particularly notable change from this is that we remove several
variations of ways to write ⁉, ❔, ‼, ❗, ❓, ❣, which all depend on
"exclamation" or "question". We've reduced the large number of outlines
for question marks and exclamation marks in the main dictionaries.

The script includes some temporary hacks to keep this commit's changes
fairly small before making amendments to individual dictionaries and
removing the misstrokes.