Use glued fingerspelled alphabet in top X dicts
Typey Type no longer uses the top X words dictionaries to supply the
word lists for top X lesson material.

This lets us replace the single-letter entries in these dictionaries
that are using fingerspelling outlines with their proper cased and glued
Plover translations.

In turn, that prevents issues like Typey Type finding "words" like "p"
when trying to create a stroke hint and awkwardly splitting a word like
"pers" e.g. treating it as the "word" "p" and suffix "^ers".

This commit doesn't touch "a" because it uses the AEU outline for the
usual, single-letter word "a".