Update "te" entry in top-10000-project-gutenberg-words.json
By using "T*/SKWRE": "te", which relies on "SKWRE": "{^e}" instead of
"T*/*E": "te" we reduce the risk of trying to fingerspell a longer word
containing "te" and the entry "T*/*E": "te" matching and causing an
unnecessary space in the middle of the word.

We also don't use "T*/*E": "{&te}" here because then the Typey Type CLI
would generate a hint of "T*/SP-S/SKWRE" by matching on "T*/SP-S": "t"
and adding a suffix, which is pointlessly verbose.